SOM-22_postUpdateDelete_Role_update

+++ board/src/main/java/com/encore/post/service/PostService.java
@@ -10,7 +10,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -83,13 +85,27 @@ public class PostService{
     }
 
     public Post update(Long id, PostReqDto postReqDto) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String email = authentication.getName(); // email 정보 꺼내기
+
         Post post = postRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("not found post"));
+        if (!post.getEmail().equals(email) && !authentication.getAuthorities().contains((new SimpleGrantedAuthority("ROLE_ADMIN")))) { // order한 사람이 아니고 admin이 아니면
+            throw new AccessDeniedException("권한이 없습니다.");
+        }
+
         post.updatePost(postReqDto.getTitle(), postReqDto.getContents());
         return post;
     }
 
     public Post delete(Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String email = authentication.getName(); // email 정보 꺼내기
+
         Post post = postRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("not found post"));
+        if (!post.getEmail().equals(email) && !authentication.getAuthorities().contains((new SimpleGrantedAuthority("ROLE_ADMIN")))) { // order한 사람이 아니고 admin이 아니면
+            throw new AccessDeniedException("권한이 없습니다.");
+        }
+
         post.deletePost();
         return post;
     }
